feat(idgen): 新增雪花 ID 生成器及 1000 分片策略支持
- 实现基于雪花算法的进程内单调递增 ID 生成器 Snowflake
- 支持按业务命名空间缓存雪花 ID 生成器实例
- 自动根据环境变量或进程信息生成稳定的 worker ID
- 新增 ShardNo 函数实现雪花 ID 对 1000 取模分片号计算
- 添加雪花 ID 生成及分片功能的单元测试代码
- 更新用户标签模块文档，明确 shard_no 固定为 uid % 1000 分片规则
- 调整用户标签运行期与结果分片相关配置，支持 1000 分片数量级
- 优化前台用户管理接口，响应 ID 改为雪花 ID，列表查询可按 shardNo 精确筛选
- 删除前台用户管理数据库命名库配置，改用默认主库绑定 user 表
- 更新管理员权限与 API 文档，增加 MFA 相关安全校验项
- 清理配置和注释，调整运行期配置加载与错误提示信息
- 修正多处文档权限码及接口路径，提高一致性和准确性

+++ internal/jobs/usertag/repository/runtime_uid.go
@@ -34,7 +34,7 @@ func (d RuntimeDeps) AddRuntimeUIDs(ctx context.Context, workflowID, scope strin
 		rows = append(rows, model.UserTagRuntimeUID{
 			WorkflowID: workflowID,
 			UID:        uid,
-			ShardNo:    int8(d.ShardPlan.UIDShard(uid, d.ShardPlan.RuntimeShardTotal)),
+			ShardNo:    d.ShardPlan.UIDShard(uid, d.ShardPlan.RuntimeShardTotal),
 			Scope:      scope,
 			CreatedAt:  now,
 			UpdatedAt:  now,
